[#3558] EmptyLineCorrector extraction
