+++ lib/rubocop.rb
@@ -140,6 +140,7 @@ require_relative 'rubocop/cop/mixin/trailing_comma'
 require_relative 'rubocop/cop/mixin/unused_argument'
 
 require_relative 'rubocop/cop/correctors/alignment_corrector'
+require_relative 'rubocop/cop/correctors/empty_line_corrector'
 
 require_relative 'rubocop/cop/bundler/duplicated_gem'
 require_relative 'rubocop/cop/bundler/insecure_protocol_source'
