+++ lib/rubocop/cop/correctors/alignment_corrector.rb
@@ -4,7 +4,7 @@ module RuboCop
   module Cop
     # This class does auto-correction of nodes that should just be moved to
     # the left or to the right, amount being determined by the instance
-    # variable @column_delta.
+    # variable column_delta.
     class AlignmentCorrector
       extend Util
