+++ lib/rubocop/cop/layout/empty_lines_around_begin_body.rb
@@ -31,6 +31,10 @@ module RuboCop
           check(node, nil)
         end
 
+        def autocorrect(node)
+          EmptyLineCorrector.correct(node)
+        end
+
         private
 
         def style
