+++ lib/rubocop/cop/layout/empty_lines_around_class_body.rb
@@ -58,6 +58,10 @@ module RuboCop
           _obj, body = *node
           check(node, body)
         end
+
+        def autocorrect(node)
+          EmptyLineCorrector.correct(node)
+        end
       end
     end
   end
