+++ lib/rubocop/cop/mixin/documentation_comment.rb
@@ -7,9 +7,9 @@ module RuboCop
       extend NodePattern::Macros
       include Style::AnnotationComment
 
-      def_node_matcher :constant_definition?, '{class module casgn}'
-
       private
+
+      def_node_matcher :constant_definition?, '{class module casgn}'
 
       def documentation_comment?(node)
         preceding_lines = preceding_lines(node)
