+++ lib/rubocop/cop/mixin/integer_node.rb
@@ -4,6 +4,8 @@ module RuboCop
   module Cop
     # Common functionality for checking integer nodes.
     module IntegerNode
+      private
+
       def integer_part(node)
         node.source.sub(/^[+-]/, '').split('.').first
       end
